day13/requirements.txt
flask
flask-socketio
flask-limiter
flask-wtf
openai
python-dotenv
tiktoken
faiss-cpu
numpy
